bug fixes
Fixing bugs that I have found.
Most are related to converting integers to pointers and vice-versa.
This had the effect that the program would work in 32 bits mode, but not
in 64 bits. Because in 32 bits mode int is the same size as a pointer.
But in 64 bits, pointer is 8 bytes long while int is only 4. This caused
pointers to be corrupted in the process of the conversion, then
the program would crash.

+++ imisc/imisc.c
@@ -385,7 +385,7 @@ imisc_error (int sd, int id, int error, struct type_IMISC_IA5List *parameter, st
 	struct RyError *rye;
 
 	if (error == RY_REJECT) {
-		advise (NULLCP, "%s", RoErrString ((int) parameter));
+		advise (NULLCP, "%s", RoErrString ((ssize_t) parameter));
 		return OK;
 	}
