+++ quipu/malloc.c
@@ -107,7 +107,7 @@ struct header {
 			unsigned short 	 control;
 			unsigned short   size;
 		} small;
-		unsigned int big;
+		size_t big;
 	} un;
 };
 
@@ -121,11 +121,11 @@ struct header {
 /* sizes chosen for anticipated QUIPU behaviour */
 
 #define BUCKETS 8
-static unsigned sizes [BUCKETS] = { 0, 12, 24, 68, 512, 1028, 8204, MAXINT};
+static size_t sizes [BUCKETS] = { 0, 12, 24, 68, 512, 1028, 8204, MAXINT};
 
 struct freelist {
 	struct header * block;
-	unsigned int size;
+	size_t size;
 	struct freelist * next;
 	struct freelist * prev;
 };
@@ -146,12 +146,12 @@ static int first_malloc = 1;
 static char * top_mem;
 
 #ifdef	BSD42
-static unsigned int pagesize = 0x2000;
+static size_t pagesize = 0x2000;
 #endif
 
 /* all calculated later - initialise for safety */
-static unsigned int pagemask = 0xE000;
-static unsigned int pageminusone = 0x1FFF;
+static size_t pagemask = 0xE000;
+static size_t pageminusone = 0x1FFF;
 static unsigned short smallmax = 0xDFFF;
 
 extern caddr_t sbrk();
@@ -269,7 +269,7 @@ write_int (unsigned x) {
 }
 
 static
-log_realloc (unsigned oldlen, unsigned newlen, unsigned bsize, char *addr) {
+log_realloc (size_t oldlen, size_t newlen, size_t bsize, char *addr) {
 	write_string ("realloc of ");
 	write_int (oldlen);
 	write_string ("at ");
@@ -288,7 +288,7 @@ log_realloc (unsigned oldlen, unsigned newlen, unsigned bsize, char *addr) {
 }
 
 static
-print_free_list (unsigned heap) {
+print_free_list (size_t heap) {
 	int i;
 	struct freelist * top;
 	struct freelist * ptr;
@@ -373,9 +373,9 @@ new_freelist (void) {
 static char *
 big_malloc (
 	/* used for mallocs of > MAXSMALL */
-	unsigned realsize
+	size_t realsize
 ) {
-	unsigned blocksize;
+	size_t blocksize;
 	struct freelist * flist;
 	struct header * head = (struct header *)0;
 	char * mem;
@@ -448,7 +448,7 @@ big_free (struct header *ptr) {
 static
 add_free (struct header *x) {
 	struct freelist *next, *c;
-	unsigned * p = sizes;
+	size_t * p = sizes;
 
 	x->use &= ~INUSE;
 
@@ -491,7 +491,7 @@ next_free_block (struct header *ptr) {
 
 	next = (struct header *)((char *)ptr + ptr->smallsize);
 
-	if (((int)(next - 1) & pagemask) != ((int)next & pagemask))
+	if (((size_t)(next - 1) & pagemask) != ((size_t)next & pagemask))
 		return (struct header *)0;
 
 	if (((char *)next < top_mem) && (next->use == (ptr->use & ~INUSE)))
@@ -516,15 +516,15 @@ x_malloc (size)
 #else
 malloc (size)
 #endif
-unsigned size;
+size_t size;
 {
 	char * mem;
 	struct header *head;
-	unsigned realsize, blocksize;
+	size_t realsize, blocksize;
 	struct freelist * top;
 	struct freelist * ptr;
 	int i;
-	unsigned * p = sizes;
+	size_t * p = sizes;
 
 	if (first_malloc) {
 #ifdef	BSD42
@@ -553,7 +553,7 @@ unsigned size;
 
 	if (first_malloc) {
 		/* set up freelist */
-		unsigned x;
+		size_t x;
 		int j;
 
 		for (i = 0; i < MAXHEAP; i++) {
@@ -567,7 +567,7 @@ unsigned size;
 		}
 
 		/* align first sbrk to page boundary */
-		x = (unsigned)sbrk(0);
+		x = (size_t)sbrk(0);
 		x = PAGEALIGN (x) - x;
 		blocksize = PAGEALIGN(realsize) + x;
 		if ((head = (struct header *) sbrk ((int)blocksize)) == (struct header *)-1) {
@@ -708,13 +708,13 @@ char *s1;
 #else
 void *s1;
 #endif
-unsigned n;
+size_t n;
 {
 	char *mem, *s = (char*) s1;
-	unsigned realsize;
+	size_t realsize;
 	struct header * ptr;
 	struct header * next;
-	unsigned copysize;
+	size_t copysize;
 
 	ptr = (struct header *) (s - ALIGN (sizeof (struct header)));
 
@@ -775,7 +775,7 @@ unsigned n;
 		/* is it big enough ? */
 		if (ptr->smallsize + top->smallsize >= realsize) {
 #ifdef MALLOCTRACE
-			unsigned savesize;
+			size_t savesize;
 			savesize = ptr->smallsize;
 #endif
 
@@ -809,10 +809,10 @@ x_calloc(n, size)
 #else
 calloc(n, size)
 #endif
-unsigned n, size;
+size_t n, size;
 {
 	char * mem;
-	unsigned x;
+	size_t x;
 
 	x= n*size;
 	if ((mem = malloc (x)) == (char *)0)
