+++ snmp/clns.c
@@ -140,7 +140,7 @@ int	offset;
 	OT	    ot = oi -> oi_type;
 	static   int lastq = -1;
 
-	ifvar = (int) ot -> ot_info;
+	ifvar = (ssize_t) ot -> ot_info;
 	switch (offset) {
 	case type_SNMP_PDUs_get__request:
 		if (oid -> oid_nelem != ot -> ot_name -> oid_nelem + 1
@@ -428,7 +428,7 @@ int	offset;
 	OT	    ot = oi -> oi_type;
 	OS	    os = ot -> ot_syntax;
 
-	ifvar = (int) ot -> ot_info;
+	ifvar = (ssize_t) ot -> ot_info;
 	switch (offset) {
 	case type_SNMP_PDUs_set__request:
 	case type_SNMP_PDUs_commit:
@@ -543,7 +543,7 @@ int	offset;
 	if (get_interfaces (offset) == NOTOK)
 		return generr (offset);
 
-	ifvar = (int) ot -> ot_info;
+	ifvar = (ssize_t) ot -> ot_info;
 	switch (offset) {
 	case type_SNMP_PDUs_get__request:
 		if (oid -> oid_nelem <= ot -> ot_name -> oid_nelem)
@@ -661,7 +661,7 @@ int	offset;
 	if (get_routes (offset) == NOTOK)
 		return generr (offset);
 
-	ifvar = (int) ot -> ot_info;
+	ifvar = (ssize_t) ot -> ot_info;
 try_again:
 	;
 	switch (offset) {
@@ -859,7 +859,7 @@ int	offset;
 	if (get_arptab (offset) == NOTOK)
 		return generr (offset);
 
-	switch (ifvar = (int) ot -> ot_info) {
+	switch (ifvar = (ssize_t) ot -> ot_info) {
 	case clnpNetToMediaIfIndex:
 	case clnpNetToMediaPhysAddress:
 	case clnpNetToMediaNetAddress:
@@ -1215,7 +1215,7 @@ int	offset;
 	OT	    ot = oi -> oi_type;
 	static   int lastq = -1;
 
-	ifvar = (int) ot -> ot_info;
+	ifvar = (ssize_t) ot -> ot_info;
 	switch (offset) {
 	case type_SNMP_PDUs_get__request:
 		if (oid -> oid_nelem != ot -> ot_name -> oid_nelem + 1
