+++ snmp/eval.c
@@ -105,7 +105,7 @@ int	offset;
 			arg2;
 	OT	    ot = oi -> oi_type;
 
-	ifvar = (int) ot -> ot_info;
+	ifvar = (ssize_t) ot -> ot_info;
 	switch (offset) {
 	case type_SNMP_PDUs_get__request:
 		if (curexpr == NULL)
@@ -267,7 +267,7 @@ int	offset;
 	OT	    ot = oi -> oi_type;
 	struct expr *e;
 
-	ifvar = (int) ot -> ot_info;
+	ifvar = (ssize_t) ot -> ot_info;
 	switch (offset) {
 	case type_SNMP_PDUs_get__request:
 		if (oid -> oid_nelem != ot -> ot_name -> oid_nelem + 1)
@@ -660,7 +660,7 @@ int	offset;
 	struct qbuf *qb;
 
 #ifndef	lint
-	ifvar = (int) ot -> ot_info;
+	ifvar = (ssize_t) ot -> ot_info;
 #endif
 	switch (offset) {
 	case type_SNMP_PDUs_set__request:
