+++ snmp/view-g.c
@@ -63,7 +63,7 @@ int	offset;
 	OID    oid = oi -> oi_name;
 	OT	    ot = oi -> oi_type;
 
-	ifvar = (int) ot -> ot_info;
+	ifvar = (ssize_t) ot -> ot_info;
 	switch (offset) {
 	case type_SNMP_PDUs_get__request:
 		if (oid -> oid_nelem <= ot -> ot_name -> oid_nelem)
@@ -214,7 +214,7 @@ int	offset;
 	OID    oid = oi -> oi_name;
 	OT	    ot = oi -> oi_type;
 
-	ifvar = (int) ot -> ot_info;
+	ifvar = (ssize_t) ot -> ot_info;
 	switch (offset) {
 	case type_SNMP_PDUs_get__request:
 		if (oid -> oid_nelem <= ot -> ot_name -> oid_nelem)
@@ -348,7 +348,7 @@ int	offset;
 	OID    oid = oi -> oi_name;
 	OT	    ot = oi -> oi_type;
 
-	ifvar = (int) ot -> ot_info;
+	ifvar = (ssize_t) ot -> ot_info;
 	switch (offset) {
 	case type_SNMP_PDUs_get__request:
 		if (oid -> oid_nelem <= ot -> ot_name -> oid_nelem)
